+++ src/main/java/com/librarymanagment/springbootlibrary/repository/DateInformationRepo.java
@@ -11,11 +11,11 @@ public interface DateInformationRepo extends JpaRepository<DateInformation,Integ
     //custom query to search for date info based on ID's
     //this function will return a DateInformation entity that it finds that matches the search
     @Query(value = "select * from date_info u where u.ISBN = :id1 and u.student_id = :id2",nativeQuery = true)
-    DateInformation searchBorrowDate(@Param("id1") Integer id1, @Param("id2") Integer id2);
+    DateInformation searchBorrowDate(@Param("id1") Long id1, @Param("id2") Integer id2);
 
     //need both Modifying and Transactional for delete operation to work
     @Modifying
     @Transactional
     @Query(value = "Delete from date_info u where u.ISBN = :id1 and u.student_id = :id2",nativeQuery = true)
-    void deleteDates(@Param("id1") Integer id1, @Param("id2") Integer id2);
+    void deleteDates(@Param("id1") Long id1, @Param("id2") Integer id2);
 }
